refactor: actually follow the format correctly for curseforge
