+++ adapters/packages/adapter-curseforge/src/Adapter/Functions/getAllVersions.ts
@@ -1,3 +1,7 @@
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
 import { ModVersion } from '@glowsquid/glowsquid-adapter'
 import { getMod } from './getMod'
 import { getVersion } from './getVersion'
