+++ adapters/packages/adapter-curseforge/src/Adapter/Functions/getMod.ts
@@ -1,6 +1,8 @@
-import type { Mod } from '@glowsquid/glowsquid-adapter'
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */import type { Mod } from '@glowsquid/glowsquid-adapter'
 import curseforge from 'mc-curseforge-api'
-import { config } from '../Adapter/Config'
+import { config } from '../Config'
 
 export const getMod = async (modId: string): Promise<Mod> => {
   const mod = await curseforge.getMod(Number(modId))
