+++ adapters/packages/adapter-curseforge/src/Adapter/Functions/getModAuthors.ts
@@ -1,3 +1,6 @@
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
 import type { Mod } from '@glowsquid/glowsquid-adapter'
 import { getMod } from './getMod'
