+++ adapters/packages/adapter-curseforge/src/Adapter/index.ts
@@ -1,0 +1,5 @@
+export * from './Functions/getMod'
+export * from './Functions/getAllVersions'
+export * from './Functions/getModList'
+export * from './Functions/getVersion'
+export * from './Functions/getModAuthors'
